Changed filename course-settings to settings
